+++ scripts/S03-TBKM.R
@@ -58,7 +58,7 @@ if(dir.exists(f)){
       file.path(expDir, paste0(db_info(toExport)$name, "-BCK-", Sys.Date()))
    ) %>% invisible()
 }
-as_fileMDB(toExport, path=expDir)
+as_fileMDB(toExport, path=expDir, htmlModel = FALSE)
 
 ###############################################################################@
 ## Save TBKM spepc ----
@@ -70,5 +70,5 @@ if(dir.exists(f)){
       file.path(expDir, paste0(db_info(toExport)$name, "-BCK-", Sys.Date()))
    ) %>% invisible()
 }
-as_fileMDB(toExport, path=expDir)
+as_fileMDB(toExport, path=expDir, htmlModel = FALSE)